Implementação do Safra na nova estrutura.
